bot/requirements.txt
matrix-nio>=0.24.0
aiohttp>=3.9.0
aiofiles>=23.0
python-slugify>=8.0
github-copilot-sdk
